+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added utility functions to store and restore tokens
+- Added a refresh token to the session table and functions to detect expired or soon to be expired tokens
+- Added global and per-request retries of failed requests
+- Added cancellation token for Rest API requests
 
 ## [3.0.3] - 2022-05-20
 ### Fixed